refactor: new path alias (@src)
