+++ src/components/core/DynamicLoading.tsx
@@ -1,6 +1,6 @@
 import dynamic from 'next/dynamic';
 
-const Loading = dynamic(() => import('@components/dynamic/Loading'), {
+const Loading = dynamic(() => import('@src/components/dynamic/Loading'), {
   loading: () => <p>loading...</p>,
 });
