+++ src/components/ui/Button.tsx
@@ -3,7 +3,7 @@ import { ElementType, Ref } from 'react';
 
 import { Props, __ } from 'types';
 
-import { forwardRefWithAs } from '@utils/forward-ref-with-as';
+import { forwardRefWithAs } from '@src/utils/forward-ref-with-as';
 
 const DEFAULT_BUTTON_TAG = 'button' as const;
