+++ src/lib/fetcher.ts
@@ -6,7 +6,7 @@ export const fetcher = ky.extend({
       async (_req, _opt, res) => {
         if (!res.ok) {
           const contentType = res.headers.get('Content-Type');
-          if (!contentType || contentType.indexOf('application/json') === -1) 
+          if (!contentType || contentType.indexOf('application/json') === -1)
             throw await res.text();
           throw await res.json();
         }
