+++ src/utils/aws/s3/index.ts
@@ -1,6 +1,6 @@
 import { S3Client } from '@aws-sdk/client-s3';
 
-import credentials from '@utils/aws/credentials';
+import credentials from '@src/utils/aws/credentials';
 
 export const s3Client = new S3Client({
   region: 'ap-northeast-2',
