+++ src/utils/mongodb/encrypt.ts
@@ -1,7 +1,7 @@
 import HashIds from 'hashids';
 import { ObjectId } from 'mongodb';
 
-import { HASHIDS_KEY } from '@defines/env';
+import { HASHIDS_KEY } from '@src/defines/env';
 
 const hashIds = new HashIds(HASHIDS_KEY);
